+++ hydownloader/db.py
@@ -492,6 +492,30 @@ def check_and_update_db() -> None:
                     log.info("hydownloader", "Updating version number...")
                     cur.execute('update version set version = \'0.19.0\'')
                 log.info("hydownloader", "Upgraded database to version 0.19.0")
+            elif version == "0.19.0": # 0.19.0 -> 0.20.0
+                log.info("hydownloader", "Starting database upgrade to version 0.20.0")
+                with sqlite3.connect(_path+"/hydownloader.db") as connection:
+                    cur = connection.cursor()
+                    cur.execute('begin exclusive transaction')
+                    write_new_config(["hydownloader-import-jobs.json", "gallery-dl-user-config.json"])
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! The default content of gallery-dl-user-config.json changed.")
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! The \"include\" key was added and set to \"avatar,background,artworks\" for \"pixiv\". This change is optional. Applying this change will cause background and artist profile images to be downloaded (in addition to artworks).")
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! See the newly written default configuration file (name ending in .NEW) in your database folder.")
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! The default content of hydownloader-import-jobs.json changed:")
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! The URL generation rules for Newgrounds were updated. Newgrounds changed their URL format, so the old rule now produces invalid post URLs.")
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! Fixes were made to the code of the \"gelbooru id\" tag rule and the URL rule that generates gelbooru post URLs (the format of the generated results did not change, but an integer->string type conversion of IDs was missing).")
+                    log.warning("hydownloader", "!!MANUAL INTERVENTION REQUIRED!! See the newly written default importer configuration file (name ending in .NEW) in your database folder.")
+                    log.info("hydownloader", "Updating version number...")
+                    cur.execute('update version set version = \'0.20.0\'')
+                log.info("hydownloader", "Upgraded database to version 0.20.0")
+            elif version == "0.20.0": # 0.20.0 -> 0.21.0
+                log.info("hydownloader", "Starting database upgrade to version 0.21.0")
+                with sqlite3.connect(_path+"/hydownloader.db") as connection:
+                    cur = connection.cursor()
+                    cur.execute('begin exclusive transaction')
+                    log.info("hydownloader", "Updating version number...")
+                    cur.execute('update version set version = \'0.21.0\'')
+                log.info("hydownloader", "Upgraded database to version 0.21.0")
             else:
                 log.fatal("hydownloader", "Unsupported hydownloader database version found")